localmodconfig: document localmodconfig in README
Someone (over a year ago :-p) asked me to document localmodconfig in the
README file in the source code.  I thought it was a good idea but other
things were more important and I simply forgot about it.  Well, I stumbled
on the email asking me about this and I'm sending it out now.

Signed-off-by: Steven "Mr. Procrastinator" Rostedt <rostedt@goodmis.org>
Cc: Michal Marek <mmarek@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>