mm: vmscan: check for fatal signals iff the process was throttled
commit 5515061d22f0 ("mm: throttle direct reclaimers if PF_MEMALLOC
reserves are low and swap is backed by network storage") introduced a
check for fatal signals after a process gets throttled for network
storage.  The intention was that if a process was throttled and got killed
that it should not trigger the OOM killer.  As pointed out by Minchan Kim
and David Rientjes, this check is in the wrong place and too broad.  If a
system is in am OOM situation and a process is exiting, it can loop in
__alloc_pages_slowpath() and calling direct reclaim in a loop.  As the
fatal signal is pending it returns 1 as if it is making forward progress
and can effectively deadlock.

This patch moves the fatal_signal_pending() check after throttling to
throttle_direct_reclaim() where it belongs.  If the process is killed
while throttled, it will return immediately without direct reclaim except
now it will have TIF_MEMDIE set and will use the PFMEMALLOC reserves.

Minchan pointed out that it may be better to direct reclaim before
returning to avoid using the reserves because there may be pages that can
easily reclaim that would avoid using the reserves.  However, we do no
such targetted reclaim and there is no guarantee that suitable pages are
available.  As it is expected that this throttling happens when
swap-over-NFS is used there is a possibility that the process will instead
swap which may allocate network buffers from the PFMEMALLOC reserves.
Hence, in the swap-over-nfs case where a process can be throtted and be
killed it can use the reserves to exit or it can potentially use reserves
to swap a few pages and then exit.  This patch takes the option of using
the reserves if necessary to allow the process exit quickly.

If this patch passes review it should be considered a -stable candidate
for 3.6.

Signed-off-by: Mel Gorman <mgorman@suse.de>
Cc: David Rientjes <rientjes@google.com>
Cc: Luigi Semenzato <semenzato@google.com>
Cc: Dan Magenheimer <dan.magenheimer@oracle.com>
Cc: KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>
Cc: Sonny Rao <sonnyrao@google.com>
Cc: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>