mm: fix calculation of dirtyable memory
The system uses global_dirtyable_memory() to calculate number of dirtyable
pages/pages that can be allocated to the page cache.  A bug causes an
underflow thus making the page count look like a big unsigned number.
This in turn confuses the dirty writeback throttling to aggressively write
back pages as they become dirty (usually 1 page at a time).  This
generally only affects systems with highmem because the underflowed count
gets subtracted from the global count of dirtyable memory.

The problem was introduced with v3.2-4896-gab8fabd

Fix is to ensure we don't get an underflowed total of either highmem
or global dirtyable memory.

Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Signed-off-by: Puneet Kumar <puneetster@chromium.org>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Tested-by: Damien Wyart <damien.wyart@free.fr>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>